test_feishu_20210822/requirements.txt
Package              Version
-------------------- ---------
Appium-Python-Client 1.2.0
atomicwrites         1.4.0
attrs                21.2.0
certifi              2021.5.30
charset-normalizer   2.0.4
colorama             0.4.4
Faker                8.12.1
idna                 3.2
iniconfig            1.1.1
packaging            21.0
pip                  21.2.4
pluggy               0.13.1
py                   1.10.0
pyparsing            2.4.7
pytest               6.2.4
python-dateutil      2.8.2
PyYAML               5.4.1
requests             2.26.0
selenium             3.141.0
setuptools           57.4.0
six                  1.16.0
text-unidecode       1.3
toml                 0.10.2
urllib3              1.26.6
